[Refactor] Move exec state reporting into ExecStateReporter
