+++ be/src/exec/pipeline/exec_state_reporter.cpp
@@ -18,11 +18,18 @@
 #include <thrift/protocol/TDebugProtocol.h>
 
 #include <memory>
+#include <sstream>
+#include <utility>
+#include <vector>
 
+#include "base/failpoint/fail_point.h"
 #include "base/network/network_util.h"
+#include "base/utility/defer_op.h"
 #include "common/config_exec_flow_fwd.h"
 #include "common/config_network_fwd.h"
 #include "common/config_rpc_client_fwd.h"
+#include "common/object_pool.h"
+#include "common/runtime_profile.h"
 #include "common/system/backend_options.h"
 #include "common/system/master_info.h"
 #include "common/util/thrift_client_cache.h"
@@ -31,9 +38,73 @@
 #include "exec/pipeline/query_context.h"
 #include "gen_cpp/FrontendService.h"
 #include "platform/thrift_rpc_helper.h"
+#include "runtime/current_thread.h"
+#include "runtime/env/global_env.h"
+#include "runtime/mem_tracker.h"
 #include "runtime/runtime_state_helper.h"
 
 namespace starrocks::pipeline {
+
+DEFINE_FAIL_POINT(report_exec_state_failed_status);
+
+namespace {
+
+RuntimeProfile* build_merged_instance_profile(QueryContext* query_ctx, FragmentContext* fragment_ctx,
+                                              ObjectPool* obj_pool) {
+    auto* instance_profile = fragment_ctx->runtime_state()->runtime_profile();
+
+    if (query_ctx->profile_level() >= TPipelineProfileLevel::type::DETAIL) {
+        return instance_profile;
+    }
+
+    RuntimeProfile* new_instance_profile = nullptr;
+    int64_t process_raw_timer = 0;
+    DeferOp defer([&new_instance_profile, &process_raw_timer]() {
+        if (new_instance_profile != nullptr) {
+            auto* process_timer = ADD_TIMER(new_instance_profile, "BackendProfileMergeTime");
+            COUNTER_SET(process_timer, process_raw_timer);
+        }
+    });
+
+    SCOPED_RAW_TIMER(&process_raw_timer);
+    std::vector<RuntimeProfile*> pipeline_profiles;
+    instance_profile->get_children(&pipeline_profiles);
+
+    std::vector<RuntimeProfile*> merged_driver_profiles;
+    for (auto* pipeline_profile : pipeline_profiles) {
+        std::vector<RuntimeProfile*> driver_profiles;
+        pipeline_profile->get_children(&driver_profiles);
+
+        if (driver_profiles.empty()) {
+            continue;
+        }
+
+        auto* merged_driver_profile = RuntimeProfile::merge_isomorphic_profiles(obj_pool, driver_profiles);
+
+        // use the name of pipeline' profile as pipeline driver's
+        merged_driver_profile->set_name(pipeline_profile->name());
+
+        // add all the info string and counters of the pipeline's profile
+        // to the pipeline driver's profile
+        merged_driver_profile->copy_all_info_strings_from(pipeline_profile);
+        merged_driver_profile->copy_all_counters_from(pipeline_profile);
+
+        merged_driver_profiles.push_back(merged_driver_profile);
+    }
+
+    new_instance_profile = obj_pool->add(new RuntimeProfile(instance_profile->name()));
+    new_instance_profile->copy_all_info_strings_from(instance_profile);
+    new_instance_profile->copy_all_counters_from(instance_profile);
+    for (auto* merged_driver_profile : merged_driver_profiles) {
+        merged_driver_profile->reset_parent();
+        new_instance_profile->add_child(merged_driver_profile, true, nullptr);
+    }
+
+    return new_instance_profile;
+}
+
+} // namespace
+
 std::string to_load_error_http_path(const std::string& file_name) {
     if (file_name.empty()) {
         return "";
@@ -173,6 +244,101 @@ Status ExecStateReporter::report_exec_status(const TReportExecStatusParams& para
     }
 
     return rpc_status;
+}
+
+void ExecStateReporter::report_exec_state(QueryContext* query_ctx, FragmentContext* fragment_ctx, const Status& status,
+                                          bool done) {
+    auto* profile = fragment_ctx->runtime_state()->runtime_profile();
+    ObjectPool obj_pool;
+    if (query_ctx->enable_profile()) {
+        profile = build_merged_instance_profile(query_ctx, fragment_ctx, &obj_pool);
+
+        // Add counters for query level memory and cpu usage, these two metrics will be specially handled at the frontend
+        auto* query_peak_memory = profile->add_counter(
+                "QueryPeakMemoryUsage", TUnit::BYTES,
+                RuntimeProfile::Counter::create_strategy(TUnit::BYTES, TCounterMergeType::SKIP_FIRST_MERGE));
+        COUNTER_SET(query_peak_memory, query_ctx->mem_cost_bytes());
+        auto* query_cumulative_cpu = profile->add_counter(
+                "QueryCumulativeCpuTime", TUnit::TIME_NS,
+                RuntimeProfile::Counter::create_strategy(TUnit::TIME_NS, TCounterMergeType::SKIP_FIRST_MERGE));
+        COUNTER_SET(query_cumulative_cpu, query_ctx->query_runtime_state().cpu_cost());
+
+        auto* query_spill_bytes = profile->add_counter(
+                "QuerySpillBytes", TUnit::BYTES,
+                RuntimeProfile::Counter::create_strategy(TUnit::BYTES, TCounterMergeType::SKIP_FIRST_MERGE));
+        COUNTER_SET(query_spill_bytes, query_ctx->get_spill_bytes());
+
+        // Add execution wall time
+        auto* query_exec_wall_time = profile->add_counter(
+                "QueryExecutionWallTime", TUnit::TIME_NS,
+                RuntimeProfile::Counter::create_strategy(TUnit::TIME_NS, TCounterMergeType::SKIP_FIRST_MERGE));
+        COUNTER_SET(query_exec_wall_time, query_ctx->lifetime());
+    }
+
+    const auto& fe_addr = fragment_ctx->fe_addr();
+    if (fe_addr.hostname.empty()) {
+        // query executed by external connectors, like spark and flink connector,
+        // does not need to report exec state to FE, so return if fe addr is empty.
+        return;
+    }
+
+    // Load channel profile will be merged on FE
+    auto* load_channel_profile = fragment_ctx->runtime_state()->load_channel_profile();
+    std::shared_ptr<TReportExecStatusParams> params;
+    {
+        // move profile memory to process, similar with SinkBuffer. the params will be released in ExecStateReporter
+        int64_t before_bytes = CurrentThread::current().get_consumed_bytes();
+        params = ExecStateReporter::create_report_exec_status_params(query_ctx, fragment_ctx, profile,
+                                                                     load_channel_profile, status, done);
+        int64_t delta = CurrentThread::current().get_consumed_bytes() - before_bytes;
+
+        CurrentThread::current().mem_release(delta);
+        GlobalEnv::GetInstance()->process_mem_tracker()->consume(delta);
+    }
+
+    auto instance_id = fragment_ctx->fragment_instance_id();
+    auto query_id = fragment_ctx->query_id();
+
+    auto report_task = [params, fe_addr, instance_id, query_id]() {
+        int retry_times = 0;
+        int max_retry_times = config::report_exec_rpc_request_retry_num;
+        while (retry_times++ < max_retry_times) {
+            auto status = ExecStateReporter::report_exec_status(*params, fe_addr);
+
+            FAIL_POINT_TRIGGER_EXECUTE(report_exec_state_failed_status, {
+                if (status.ok()) {
+                    status = Status::InternalError("injected failed status");
+                }
+            });
+
+            if (!status.ok()) {
+                if (status.is_not_found()) {
+                    VLOG(1) << "[Driver] Fail to report exec state due to query not found. fragment_instance_id="
+                            << print_id(instance_id) << ", query_id=" << print_id(query_id);
+                } else {
+                    LOG(WARNING) << "[Driver] Fail to report exec state. fragment_instance_id=" << print_id(instance_id)
+                                 << ", query_id=" << print_id(query_id) << ", status: " << status.to_string()
+                                 << ", retry_times=" << retry_times << ", max_retry_times=" << max_retry_times;
+                    // if it is done exec state report, we should retry
+                    if (params->__isset.done && params->done) {
+                        continue;
+                    }
+                }
+            } else {
+                VLOG(1) << "[Driver] Succeed to report exec state. fragment_instance_id=" << print_id(instance_id)
+                        << ", query_id=" << print_id(query_id) << ", is_done=" << params->done;
+            }
+            break;
+        }
+    };
+
+    // if it is done exec state report, We need to ensure that this report is executed with priority
+    // and is retried as much as possible to ensure success.
+    // Otherwise, it may result in the ingestion status getting stuck.
+    bool priority = done && fragment_ctx->runtime_state()->query_options().query_type == TQueryType::LOAD;
+    submit(std::move(report_task), priority);
+    VLOG(2) << "[Driver] Submit exec state report task. fragment_instance_id=" << print_id(instance_id)
+            << ", query_id=" << print_id(query_id) << ", is_done=" << done;
 }
 
 ExecStateReporter::ExecStateReporter(const CpuUtil::CpuIds& cpuids, ExecStateReporterMetrics* metrics) {
