+++ be/src/exec/pipeline/exec_state_reporter.h
@@ -14,6 +14,7 @@
 
 #pragma once
 
+#include <functional>
 #include <memory>
 
 #include "common/status.h"
@@ -39,6 +40,8 @@ public:
             RuntimeProfile* load_channel_profile, const Status& status, bool done);
 
     static Status report_exec_status(const TReportExecStatusParams& params, const TNetworkAddress& fe_addr);
+
+    void report_exec_state(QueryContext* query_ctx, FragmentContext* fragment_ctx, const Status& status, bool done);
 
     void submit(std::function<void()>&& report_task, bool priority = false);
